+++ example/android/app/src/main/kotlin/com/iw/mint/flutter_demo/Demo.kt
@@ -1,0 +1,10 @@
+package com.iw.mint.flutter_demo
+
+import investwell.activity.AppApplication
+
+
+class Demo : AppApplication(){
+    override fun onCreate() {
+        super.onCreate()
+    }
+}